requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
attrs==23.2.0
auditd-tools==0.0.3
certifi==2024.6.2
charset-normalizer==3.3.2
frozenlist==1.4.1
geoip2==4.8.0
gitdb==4.0.11
GitPython==3.1.41
idna==3.7
maxminddb==2.6.2
multidict==6.0.5
python-dotenv==1.0.1
requests==2.32.3
setuptools==69.0.3
smmap==5.0.1
urllib3==2.2.2
wheel==0.43.0
yarl==1.9.4
